requirements/base.txt
asgiref>=3.7.2,<3.8
Django>=4.2.5,<4.3
python-dotenv>=1.0.0,<1.1
sqlparse>=0.4.4,<0.5
tzdata>=2023.3,<2023.4
redis>=5.0.0,<5.1
djangorestframework>=3.14.0,<3.15
